Move plan modifiers to separate packages
